Downgrade aws-sdk to aws-sdk-v1
Unfortunately AWS API v1 is not compatible
with AP v2 whcih results in a lot a wired
problems.

Mainly `uninitialized constant Aws`. So it
seems v1 is used because the namespace for
v1 is `AWS`. In general bundler should help
to fix such problems but when using this
library in a LWRP for chef things get ugly.